move priceCalculator and StockPriceCalculator - to stock.domain.calculator package
