+++ src/main/java/co/fineants/stock/application/SearchStock.java
@@ -15,9 +15,9 @@ import co.fineants.api.domain.kis.service.ClosingPriceService;
 import co.fineants.api.domain.kis.service.CurrentPriceService;
 import co.fineants.api.global.common.time.LocalDateTimeService;
 import co.fineants.api.global.errors.exception.business.StockNotFoundException;
-import co.fineants.stock.domain.PriceCalculator;
 import co.fineants.stock.domain.Stock;
 import co.fineants.stock.domain.StockRepository;
+import co.fineants.stock.domain.calculator.PriceCalculator;
 import co.fineants.stock.infrastructure.StockQueryDslRepository;
 import co.fineants.stock.presentation.dto.response.StockResponse;
 import co.fineants.stock.presentation.dto.response.StockSearchItem;
