+++ src/main/java/co/fineants/stock/domain/calculator/PriceCalculator.java
@@ -1,4 +1,4 @@
-package co.fineants.stock.domain;
+package co.fineants.stock.domain.calculator;
 
 import co.fineants.api.domain.common.money.Expression;
